Validate Authentication while profile update

+++ app/Http/Livewire/User/Settings/Profile.php
@@ -63,33 +63,37 @@ class Profile extends Component
     public function updateProfile()
     {
         if (Auth::check()) {
-            $this->validate([
-                'firstname' => 'nullable|max:30',
-                'lastname' => 'nullable|max:30',
-                'bio' => 'nullable|max:1000',
-                'location' => 'nullable|max:30',
-                'company' => 'nullable|max:30',
-                'avatar' => 'nullable|mimes:jpeg,jpg,png,gif|max:2048',
-            ]);
-
-            if ($this->avatar) {
-                $old_avatar = explode('storage/', $this->user->avatar);
-                if (array_key_exists(1, $old_avatar)) {
-                    Storage::delete($old_avatar[1]);
+            if (Auth::id() === $this->user->id) {
+                $this->validate([
+                    'firstname' => 'nullable|max:30',
+                    'lastname' => 'nullable|max:30',
+                    'bio' => 'nullable|max:1000',
+                    'location' => 'nullable|max:30',
+                    'company' => 'nullable|max:30',
+                    'avatar' => 'nullable|mimes:jpeg,jpg,png,gif|max:2048',
+                ]);
+    
+                if ($this->avatar) {
+                    $old_avatar = explode('storage/', $this->user->avatar);
+                    if (array_key_exists(1, $old_avatar)) {
+                        Storage::delete($old_avatar[1]);
+                    }
+                    $avatar = $this->avatar->store('avatars');
+                    $this->user->avatar = config('app.url').'/storage/'.$avatar;
                 }
-                $avatar = $this->avatar->store('avatars');
-                $this->user->avatar = config('app.url').'/storage/'.$avatar;
-            }
-
-            if (Auth::check()) {
-                $this->user->firstname = $this->firstname;
-                $this->user->lastname = $this->lastname;
-                $this->user->bio = $this->bio;
-                $this->user->location = $this->location;
-                $this->user->company = $this->company;
-                $this->user->save();
-
-                return session()->flash('profile', 'Your profile has been updated!');
+    
+                if (Auth::check()) {
+                    $this->user->firstname = $this->firstname;
+                    $this->user->lastname = $this->lastname;
+                    $this->user->bio = $this->bio;
+                    $this->user->location = $this->location;
+                    $this->user->company = $this->company;
+                    $this->user->save();
+    
+                    return session()->flash('profile', 'Your profile has been updated!');
+                }
+            } else {
+                return session()->flash('error', 'Forbidden!');
             }
         } else {
             return session()->flash('error', 'Forbidden!');
@@ -99,14 +103,18 @@ class Profile extends Component
     public function useGravatar()
     {
         if (Auth::check()) {
-            $old_avatar = explode('storage/', $this->user->avatar);
-            if (array_key_exists(1, $old_avatar)) {
-                Storage::delete($old_avatar[1]);
+            if (Auth::id() === $this->user->id) {
+                $old_avatar = explode('storage/', $this->user->avatar);
+                if (array_key_exists(1, $old_avatar)) {
+                    Storage::delete($old_avatar[1]);
+                }
+                $this->user->avatar = 'https://secure.gravatar.com/avatar/'.md5(Auth::user()->email).'?s=500&d=identicon';
+                $this->user->save();
+    
+                return session()->flash('profile', 'Your profile has been updated!');
+            } else {
+                return session()->flash('error', 'Forbidden!');
             }
-            $this->user->avatar = 'https://secure.gravatar.com/avatar/'.md5(Auth::user()->email).'?s=500&d=identicon';
-            $this->user->save();
-
-            return session()->flash('profile', 'Your profile has been updated!');
         } else {
             return session()->flash('error', 'Forbidden!');
         }
@@ -115,15 +123,19 @@ class Profile extends Component
     public function updateSponsor()
     {
         if (Auth::check()) {
-            $this->validate([
-                'sponsor' => 'nullable|active_url',
-            ]);
-
-            if (Auth::check()) {
-                $this->user->sponsor = $this->sponsor;
-                $this->user->save();
-
-                return session()->flash('sponsor', 'Your sponsor link has been updated!');
+            if (Auth::id() === $this->user->id) {
+                $this->validate([
+                    'sponsor' => 'nullable|active_url',
+                ]);
+    
+                if (Auth::check()) {
+                    $this->user->sponsor = $this->sponsor;
+                    $this->user->save();
+    
+                    return session()->flash('sponsor', 'Your sponsor link has been updated!');
+                }
+            } else {
+                return session()->flash('error', 'Forbidden!');
             }
         } else {
             return session()->flash('error', 'Forbidden!');
@@ -133,25 +145,29 @@ class Profile extends Component
     public function updateSocial()
     {
         if (Auth::check()) {
-            $this->validate([
-                'website' => 'nullable|active_url',
-                'twitter' => 'nullable|alpha_dash|max:30',
-                'twitch' => 'nullable|alpha_dash|max:200',
-                'telegram' => 'nullable|alpha_dash|max:30',
-                'github' => 'nullable|alpha_dash|max:30',
-                'youtube' => 'nullable|alpha_dash|max:30',
-            ]);
-
-            if (Auth::check()) {
-                $this->user->website = $this->website;
-                $this->user->twitter = $this->twitter;
-                $this->user->twitch = $this->twitch;
-                $this->user->telegram = $this->telegram;
-                $this->user->github = $this->github;
-                $this->user->youtube = $this->youtube;
-                $this->user->save();
-
-                return session()->flash('social', 'Your social links has been updated!');
+            if (Auth::id() === $this->user->id) {
+                $this->validate([
+                    'website' => 'nullable|active_url',
+                    'twitter' => 'nullable|alpha_dash|max:30',
+                    'twitch' => 'nullable|alpha_dash|max:200',
+                    'telegram' => 'nullable|alpha_dash|max:30',
+                    'github' => 'nullable|alpha_dash|max:30',
+                    'youtube' => 'nullable|alpha_dash|max:30',
+                ]);
+    
+                if (Auth::check()) {
+                    $this->user->website = $this->website;
+                    $this->user->twitter = $this->twitter;
+                    $this->user->twitch = $this->twitch;
+                    $this->user->telegram = $this->telegram;
+                    $this->user->github = $this->github;
+                    $this->user->youtube = $this->youtube;
+                    $this->user->save();
+    
+                    return session()->flash('social', 'Your social links has been updated!');
+                }
+            } else {
+                return session()->flash('error', 'Forbidden!');
             }
         } else {
             return session()->flash('error', 'Forbidden!');
@@ -161,7 +177,7 @@ class Profile extends Component
     public function onlyFollowingsTasks()
     {
         if (Auth::check()) {
-            if (Auth::check() && Auth::id() === $this->user->id) {
+            if (Auth::id() === $this->user->id) {
                 $this->user->onlyFollowingsTasks = ! $this->user->onlyFollowingsTasks;
                 $this->user->save();
                 if ($this->user->onlyFollowingsTasks) {
@@ -170,7 +186,7 @@ class Profile extends Component
                     session()->flash('showfollowing', 'All user\'s task will be show on homepage');
                 }
             } else {
-                return false;
+                return session()->flash('error', 'Forbidden!');
             }
         } else {
             return session()->flash('error', 'Forbidden!');
